requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.25
psycopg[binary]==3.1.18
geoalchemy2==0.14.3
alembic==1.13.1
influxdb-client==1.39.0
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
geopy==2.4.1
shapely==2.0.2

# Machine Learning
tensorflow==2.15.0
scikit-learn==1.4.0
joblib==1.3.2

# Routing & Maps
osrm-py==0.1.0
requests==2.31.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0
